+++ extensions/ql-vscode/src/view/data-flow-paths/DataFlowPathsView.tsx
@@ -1,18 +1,19 @@
 import * as React from "react";
 import { useEffect, useState } from "react";
 import { ToDataFlowPathsMessage } from "../../pure/interface-types";
-import { DataFlowPaths } from "../../variant-analysis/shared/data-flow-paths";
+import { DataFlowPaths as DataFlowPathsDomainModel } from "../../variant-analysis/shared/data-flow-paths";
+import { DataFlowPaths } from "./DataFlowPaths";
 
 export type DataFlowPathsViewProps = {
-  dataFlowPaths?: DataFlowPaths;
+  dataFlowPaths?: DataFlowPathsDomainModel;
 };
 
 export function DataFlowPathsView({
   dataFlowPaths: initialDataFlowPaths,
 }: DataFlowPathsViewProps): JSX.Element {
-  const [dataFlowPaths, setDataFlowPaths] = useState<DataFlowPaths | undefined>(
-    initialDataFlowPaths,
-  );
+  const [dataFlowPaths, setDataFlowPaths] = useState<
+    DataFlowPathsDomainModel | undefined
+  >(initialDataFlowPaths);
 
   useEffect(() => {
     const listener = (evt: MessageEvent) => {
@@ -20,6 +21,10 @@ export function DataFlowPathsView({
         const msg: ToDataFlowPathsMessage = evt.data;
         if (msg.t === "setDataFlowPaths") {
           setDataFlowPaths(msg.dataFlowPaths);
+
+          // Scroll to the top of the page when we're rendering
+          // new data flow paths.
+          window.scrollTo(0, 0);
         }
       } else {
         // sanitize origin
@@ -38,11 +43,5 @@ export function DataFlowPathsView({
     return <>Loading data flow paths</>;
   }
 
-  // For now, just render the data flows as JSON.
-  return (
-    <>
-      Loaded
-      <pre>{JSON.stringify(dataFlowPaths)}</pre>
-    </>
-  );
+  return <DataFlowPaths dataFlowPaths={dataFlowPaths} />;
 }